+++ stdlib/public/SDK/AVFoundation/NSValue.swift.gyb
@@ -1,0 +1,29 @@
+@_exported import AVFoundation // Clang module
+import CoreMedia
+import Foundation
+
+%{
+from gyb_foundation_support import \
+  ObjectiveCBridgeableImplementationForNSValueWithCategoryMethods
+}%
+
+// Bridge CoreMedia structs to NSValue.
+// AVFoundation provides internal NSValue subclasses for these structures that
+// are incompatible with the NSConcreteValue subclasses you get using
+// -[NSValue valueWithBytes:objCType:].
+
+${ ObjectiveCBridgeableImplementationForNSValueWithCategoryMethods(
+  Type="CMTime",
+  initializer="{ NSValue(time: $0) }",
+  getter="{ $0.timeValue }",
+) }
+${ ObjectiveCBridgeableImplementationForNSValueWithCategoryMethods(
+  Type="CMTimeRange",
+  initializer="{ NSValue(timeRange: $0) }",
+  getter="{ $0.timeRangeValue }",
+) }
+${ ObjectiveCBridgeableImplementationForNSValueWithCategoryMethods(
+  Type="CMTimeMapping",
+  initializer="{ NSValue(timeMapping: $0) }",
+  getter="{ $0.timeMappingValue }",
+) }
