+++ stdlib/public/SDK/AppKit/NSError.swift
@@ -1,3 +1,16 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2016 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+@_exported import AppKit
+
 extension CocoaError.Code {
   public static var textReadInapplicableDocumentType: CocoaError.Code {
     return CocoaError.Code(rawValue: 65806)
