+++ stdlib/public/SDK/CoreLocation/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_swift_library(swiftCoreLocation ${SWIFT_SDK_OVERLAY_LIBRARY_BUILD_TYPES} IS_SDK_OVERLAY
-  CoreLocation.swift.gyb
+  CLError.swift
+  NSValue.swift.gyb
+
   SWIFT_COMPILE_FLAGS "${SWIFT_RUNTIME_SWIFT_COMPILE_FLAGS}"
   LINK_FLAGS "${SWIFT_RUNTIME_SWIFT_LINK_FLAGS}"
   SWIFT_MODULE_DEPENDS Foundation
