+++ stdlib/public/SDK/CoreLocation/NSValue.swift.gyb
@@ -1,0 +1,8 @@
+@_exported import CoreLocation
+import Foundation
+
+%{
+from gyb_foundation_support import ObjectiveCBridgeableImplementationForNSValue
+}%
+
+${ ObjectiveCBridgeableImplementationForNSValue("CLLocationCoordinate2D") }
