+++ stdlib/public/SDK/CoreMedia/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_swift_library(swiftCoreMedia ${SWIFT_SDK_OVERLAY_LIBRARY_BUILD_TYPES} IS_SDK_OVERLAY
-  CoreMedia.swift
   CMTime.swift
   CMTimeRange.swift
