+++ stdlib/public/SDK/Intents/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_swift_library(swiftIntents ${SWIFT_SDK_OVERLAY_LIBRARY_BUILD_TYPES} IS_SDK_OVERLAY
-  Intents.swift
   INBooleanResolutionResult.swift
   INDoubleResolutionResult.swift
   INIntegerResolutionResult.swift
