+++ stdlib/public/SDK/MapKit/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_swift_library(swiftMapKit ${SWIFT_SDK_OVERLAY_LIBRARY_BUILD_TYPES} IS_SDK_OVERLAY
-  MapKit.swift.gyb
+  NSValue.swift.gyb
   SWIFT_COMPILE_FLAGS "${SWIFT_RUNTIME_SWIFT_COMPILE_FLAGS}"
   LINK_FLAGS "${SWIFT_RUNTIME_SWIFT_LINK_FLAGS}"
   SWIFT_MODULE_DEPENDS ObjectiveC Foundation CoreLocation
