+++ stdlib/public/SDK/Photos/PHChange.swift
