+++ stdlib/public/SDK/os/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_swift_library(swiftos ${SWIFT_SDK_OVERLAY_LIBRARY_BUILD_TYPES} IS_SDK_OVERLAY
-  os.swift
+  os_log.swift
   os.mm
 
   SWIFT_COMPILE_FLAGS "${SWIFT_RUNTIME_SWIFT_COMPILE_FLAGS}"
